common, xe: consolidate serialization API
