+++ src/common/cache_blob_id.hpp
@@ -1,5 +1,5 @@
 /*******************************************************************************
-* Copyright 2021 Intel Corporation
+* Copyright 2021-2025 Intel Corporation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -23,7 +23,7 @@
 #include <vector>
 #include <type_traits>
 
-#include "common/serialization_stream.hpp"
+#include "common/serialization.hpp"
 
 namespace dnnl {
 namespace impl {
