+++ src/gpu/intel/jit/conv/model_bridge.cpp
@@ -193,7 +193,7 @@ gradient_boost_regressor_t &get_gbr(const conv_config_t &cfg) {
         for (auto &kv : kind2data) {
             auto kind = kv.first;
             auto &data = *kv.second;
-            auto s = serialized_t::from_data(unpack_data(data));
+            auto s = serialization_stream_t::from_data(unpack_data(data));
             deserializer_t d(s);
             gbr_map[kind] = gradient_boost_regressor_t::deserialize(d);
         }
