+++ src/gpu/intel/jit/conv/zero_out.hpp
@@ -49,8 +49,8 @@ public:
             gpu_primitive_t *primitive, impl::engine_t *engine) const override;
     status_t create_generator(const compute::compute_engine_t &engine,
             compute::kernel_t &kernel) const;
-    serialized_t serialize() const override;
-    static zero_out_kernel_desc_t deserialize(const serialized_t &s);
+    serialization_stream_t serialize() const override;
+    static zero_out_kernel_desc_t deserialize(const serialization_stream_t &s);
 
     static compute::nd_range_t nd_range(int simd, size_t size);
