+++ src/gpu/intel/jit/ir/kernel_desc.hpp
@@ -17,11 +17,11 @@
 #ifndef GPU_INTEL_JIT_IR_KERNEL_DESC_HPP
 #define GPU_INTEL_JIT_IR_KERNEL_DESC_HPP
 
+#include "common/serialization.hpp"
 #include "gpu/intel/compute/compute_engine.hpp"
 #include "gpu/intel/compute/utils.hpp"
 #include "gpu/intel/jit/ir/fma.hpp"
 #include "gpu/intel/jit/ir/hw.hpp"
-#include "gpu/intel/serialization.hpp"
 
 namespace dnnl {
 namespace impl {
@@ -53,7 +53,7 @@ public:
             const impl::engine_t *engine) const = 0;
     virtual status_t create_kernel(compute::kernel_t &kernel,
             gpu_primitive_t *primitive, impl::engine_t *engine) const = 0;
-    virtual serialized_t serialize() const = 0;
+    virtual serialization_stream_t serialize() const = 0;
 };
 
 class kernel_params_base_t {
