+++ src/gpu/intel/jit/ir/kernel_info.hpp
@@ -24,9 +24,9 @@
 
 #include "common/c_types_map.hpp"
 #include "common/primitive_exec_types.hpp"
+#include "common/serialization.hpp"
 #include "gpu/intel/gpu_primitive.hpp"
 #include "gpu/intel/jit/ir/kernel_desc.hpp"
-#include "gpu/intel/serialization.hpp"
 
 namespace dnnl {
 namespace impl {
