+++ src/gpu/intel/jit/v2/conv/model.hpp
@@ -17,9 +17,9 @@
 #ifndef GPU_INTEL_JIT_V2_CONV_MODEL_HPP
 #define GPU_INTEL_JIT_V2_CONV_MODEL_HPP
 
+#include "common/serialization.hpp"
 #include "gpu/intel/jit/v2/conv/bench_data.hpp"
 #include "gpu/intel/jit/v2/conv/problem.hpp"
-#include "gpu/intel/serialization.hpp"
 
 namespace dnnl {
 namespace impl {
