+++ src/gpu/intel/ocl/bnorm/reusable_bnorm.hpp
@@ -21,6 +21,7 @@
 
 #include "common/c_types_map.hpp"
 #include "common/primitive.hpp"
+#include "common/serialization.hpp"
 #include "gpu/gpu_batch_normalization_pd.hpp"
 #include "gpu/gpu_resource.hpp"
 #include "gpu/intel/compute/dispatch_reusable.hpp"
@@ -29,7 +30,6 @@
 #include "gpu/intel/ocl/stream.hpp"
 #include "gpu/intel/ocl/utils.hpp"
 #include "gpu/intel/primitive_conf.hpp"
-#include "gpu/intel/serialization.hpp"
 
 namespace dnnl {
 namespace impl {
@@ -57,12 +57,13 @@ struct reusable_bnorm_params_t {
 #if __cplusplus >= 202002L
     bool operator==(const reusable_bnorm_params_t &) const = default;
 #endif
-    serialized_t serialize() const {
-        assert_trivially_serializable(reusable_bnorm_params_t);
-        return serialized_t(*this);
+    serialization_stream_t serialize() const {
+        DNNL_ASSERT_TRIVIALLY_SERIALIZABLE(reusable_bnorm_params_t);
+        return serialization_stream_t(*this);
     }
 
-    static reusable_bnorm_params_t deserialize(const serialized_t &s) {
+    static reusable_bnorm_params_t deserialize(
+            const serialization_stream_t &s) {
         reusable_bnorm_params_t t {};
         deserializer_t d(s);
         d.pop(t);
