+++ src/gpu/intel/ocl/gemm/ref_gemm.hpp
@@ -17,11 +17,11 @@
 #ifndef GPU_INTEL_OCL_GEMM_REF_GEMM_HPP
 #define GPU_INTEL_OCL_GEMM_REF_GEMM_HPP
 
+#include "common/serialization.hpp"
 #include "gpu/gpu_gemm_pd.hpp"
 #include "gpu/gpu_resource.hpp"
 #include "gpu/intel/gemm/gpu_gemm.hpp"
 #include "gpu/intel/primitive_conf.hpp"
-#include "gpu/intel/serialization.hpp"
 
 namespace dnnl {
 namespace impl {
