+++ src/gpu/intel/ocl/reduction/reusable_ref_reduction.hpp
@@ -1,5 +1,5 @@
 /*******************************************************************************
-* Copyright 2023-2024 Intel Corporation
+* Copyright 2023-2025 Intel Corporation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -19,13 +19,13 @@
 
 #include "common/c_types_map.hpp"
 #include "common/primitive.hpp"
+#include "common/serialization.hpp"
 #include "gpu/gpu_reduction_pd.hpp"
 #include "gpu/intel/compute/device_info.hpp"
 #include "gpu/intel/compute/dispatch_reusable.hpp"
 #include "gpu/intel/gpu_primitive.hpp"
 #include "gpu/intel/ocl/reduction/reduction_utils.hpp"
 #include "gpu/intel/primitive_conf.hpp"
-#include "gpu/intel/serialization.hpp"
 
 namespace dnnl {
 namespace impl {
@@ -58,7 +58,7 @@ struct ref_reduction_key_params_t
 
     compute::dispatch_compile_params_t params;
 };
-assert_trivially_serializable(ref_reduction_key_params_t);
+DNNL_ASSERT_TRIVIALLY_SERIALIZABLE(ref_reduction_key_params_t);
 
 struct ref_reduction_conf_t {
     ref_reduction_conf_t(const reduction_subproblem_t &subprb,
