+++ src/gpu/intel/ocl/reusable_vectorized_lnorm.hpp
@@ -19,12 +19,12 @@
 
 #include "common/c_types_map.hpp"
 #include "common/layer_normalization_pd.hpp"
+#include "common/serialization.hpp"
 #include "common/utils.hpp"
 #include "gpu/gpu_layer_normalization_pd.hpp"
 #include "gpu/intel/compute/dispatch_reusable.hpp"
 #include "gpu/intel/compute/kernel_ctx.hpp"
 #include "gpu/intel/gpu_primitive.hpp"
-#include "gpu/intel/serialization.hpp"
 
 namespace dnnl {
 namespace impl {
